requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
sqlalchemy==2.0.44
pymysql==1.0.3
pandas==2.2.2
plotly==5.15.0
matplotlib==3.8.1
openpyxl==3.1.2
xlrd==2.0.1
python-dotenv==1.1.1
gunicorn==21.2.0
WTForms==3.0.1
email-validator==2.1.0